Enforce `double` type in the download progress path
Wasn't sure where to exactly put this, or whether to split it, but it's very small change to worry about, so I guess it's fine being here

+++ osu.Game/Online/DownloadTrackingComposite.cs
@@ -144,7 +144,7 @@ namespace osu.Game.Online
 
         private void onRequestSuccess(string _) => Schedule(() => State.Value = DownloadState.Importing);
 
-        private void onRequestProgress(float progress) => Schedule(() => Progress.Value = progress);
+        private void onRequestProgress(double progress) => Schedule(() => Progress.Value = progress);
 
         private void onRequestFailure(Exception e) => Schedule(() => attachDownload(null));
